+++ include/shape.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   shape.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mkhan <mkhan@student.42.fr>                +#+  +:+       +#+        */
+/*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/25 23:07:55 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/10 14:45:16 by mkhan            ###   ########.fr       */
+/*   Updated: 2023/01/10 15:42:31 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,14 +43,14 @@ typedef struct s_shape_props	t_props;
 struct s_shape_props
 {
 	bool			highlighted;
-	double			radius;
-	double			radius_squared;
-	double			height;
-	double			diffuse;
-	double			specular;
-	double			shininess;
-	double			reflectiveness;
-	double			distance_from_origin;
+	float			radius;
+	float			radius_squared;
+	float			height;
+	float			diffuse;
+	float			specular;
+	float			shininess;
+	float			reflectiveness;
+	float			distance_from_origin;
 	t_color			color;
 	t_vector		rot;
 	t_vector		scale;
