+++ src/intersections/normal.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   normal.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mkhan <mkhan@student.42.fr>                +#+  +:+       +#+        */
+/*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/02 17:52:03 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/10 14:45:16 by mkhan            ###   ########.fr       */
+/*   Updated: 2023/01/10 15:42:31 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 t_vector	cylinder_normal(const t_shape *shape, const t_vector *itx_point)
 {
-	double		distance;
+	float		distance;
 	t_vector	normal;
 	t_vector	point;
 	t_vector	world_normal;
@@ -41,7 +41,7 @@ t_vector	cylinder_normal(const t_shape *shape, const t_vector *itx_point)
 
 t_vector	cone_normal(const t_shape *shape, const t_vector *itx_point)
 {
-	double		distance;
+	float		distance;
 	t_vector	normal;
 	t_vector	point;
 	t_vector	world_normal;
@@ -85,7 +85,7 @@ t_vector	plane_normal(const t_shape *shape)
 
 t_vector	cube_normal(const t_shape *shape, const t_vector *itx_point)
 {
-	double		maxc;
+	float		maxc;
 	t_vector	normal;
 
 	mat_vec_multiply(&normal, &shape->inv_transf, itx_point);
