+++ src/math/vector_arithmetic.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   vector_arithmetic.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mkhan <mkhan@student.42.fr>                +#+  +:+       +#+        */
+/*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 09:46:11 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/10 14:45:16 by mkhan            ###   ########.fr       */
+/*   Updated: 2023/01/10 15:42:31 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ void	sub_vec(t_vector *res, const t_vector *v1, const t_vector *v2)
  * @param v Vector to be scaled
  * @param scale scaling factor
  */
-void	scale_vec(t_vector *res, const t_vector *v, double scale)
+void	scale_vec(t_vector *res, const t_vector *v, float scale)
 {
 	res->x = scale * v->x;
 	res->y = scale * v->y;
